Begin refactoring out doc generator template
